renamed components to homepage and activity
